requirements.txt
# DailyGameLauncher 依存ライブラリ
# インストール方法: pip install -r requirements.txt

PyQt6>=6.4.0
pystray>=0.19.0
Pillow>=9.0.0
psutil>=5.9.0
keyboard>=0.13.5
requests>=2.28.0
